Add simple syntax highlighting to the JSON input box.
There are still a lot of problems to solve with this method of syntax
highlighting. The main problem being that when the syntax is updated it
will move the cursor back to the beginning as the DOM has been created
from scratch. A better method would be to insert the highlighting as
children of the existing parent elements.